move route utility function into one file
